Mildly fixed Maintain issue.
Description:
Had to rebuild database because of bug in maintain.py. However, the good news is that this time when I ran the script, it didn't delete most of the entries in the database, it just deleted three columns and didn't provide the reason or actions taken in the report.

v0.0.1